# specs/001-ai-consultation-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Car Service AI Assistant API
  version: 0.5.0
  description: >
    Полный операционный контракт, сгенерированный из runtime Express
    (`docs/api-route-inventory.json`). Схемы тел запросов — минимальные;
    источник истины по наличию path+method — инвентарь + `npm run openapi:check`.
servers:
  - url: http://localhost:3000/api
...
paths:
...
  "/analytics/ai-feedback":
    get:
      tags: [Analytics]
      summary: Получить ai-feedback
      operationId: get_analytics_ai_feedback
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/analytics/ai-feedback.csv":
    get:
      tags: [Analytics]
      summary: Получить ai-feedback.csv
      operationId: get_analytics_ai_feedback_csv
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/analytics/kpi":
    get:
      tags: [Analytics]
      summary: Получить kpi
      operationId: get_analytics_kpi
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/analytics/kpi.csv":
    get:
      tags: [Analytics]
      summary: Получить kpi.csv
      operationId: get_analytics_kpi_csv
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/analytics/manager-kpi":
    get:
      tags: [Analytics]
      summary: Получить manager-kpi
      operationId: get_analytics_manager_kpi
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/analytics/summary":
    get:
      tags: [Analytics]
      summary: Получить summary
      operationId: get_analytics_summary
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  "/bookings":
    get:
      tags: [Bookings]
      summary: Получить bookings
      operationId: get_bookings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingList' }
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags: [Bookings]
      summary: Создать/выполнить bookings
      operationId: post_bookings
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookingCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/bookings/guest":
    post:
      tags: [Bookings]
      summary: Создать/выполнить guest
      operationId: post_bookings_guest
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookingCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/bookings/{bookingId}":
    get:
      tags: [Bookings]
      summary: "Получить {bookingId}"
      operationId: get_bookings_bookingId
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      tags: [Bookings]
      summary: "Обновить {bookingId}"
      operationId: patch_bookings_bookingId
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/bookings/{bookingId}/audit":
    get:
      tags: [Bookings]
      summary: Получить audit
      operationId: get_bookings_bookingId_audit
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        error: { type: string }
        code: { type: string }
...
    BookingCreate:
      type: object
      required: [preferredAt]
      properties:
        preferredAt: { type: string, format: date-time }
        notes: { type: string }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string }
        preferredAt: { type: string, format: date-time }
    BookingList:
      oneOf:
        - type: array
          items: { $ref: '#/components/schemas/Booking' }
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Booking' }
            nextCursor: { type: string, nullable: true }
...